Extract dynamic file import to own file and initialize while creating jayvee services
