+++ libs/language-server/src/lib/workspace/index.ts
@@ -3,4 +3,5 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 export { getStdLib } from './stdlib';
+export { addDynamicFileImport } from './dynamic-file-import';
 export * from './jayvee-workspace-manager';
